+++ src/extension.ts
@@ -1,15 +1,13 @@
 import * as vscode from 'vscode';
-import copyFileName from './copyFileName';
-import copyFolderStructure from './copyStructure';
-import createProjectFromStructure from './createStructure';
+import { registerCommands } from './commands';
 
 export function activate(context: vscode.ExtensionContext): void {
-	context.subscriptions.push(copyFolderStructure);
-	context.subscriptions.push(createProjectFromStructure);
-	context.subscriptions.push(copyFileName);
+	registerCommands(context);
 }
 
+export function deactivate(): void { }
+
 /*
- * Copyright (c) 2024 Shrey Purohit.
+ * Copyright (c) 2025 Shrey Purohit.
  * This code is licensed under the MIT License.
  */