+++ CHANGELOG.md
@@ -4,5 +4,9 @@
 - v0.2 Added Banner and Extension Icon
 - v0.3 Added Major Support `Create Folder Structure`
 - v0.4 Added Feature `Copy File Name`
+- v0.5 Changed `Plain Text`
+    - (New) GitIngest Format
+    - (New) Example Preview
+    - Updated Preview Mode
 
 ## [Unreleased]